repackage appointments and refactror AppointmentRepo
